fix, refactor. Sprint#13 . Iteration 2. Updated according reviewer comments.
