+++ src/main/java/ru/practicum/shareit/item/controller/ItemController.java
@@ -4,10 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.exception.MyValidationException;
 import ru.practicum.shareit.item.dto.ItemDto;
-import ru.practicum.shareit.item.model.Item;
-import ru.practicum.shareit.item.service.ItemService;
+import ru.practicum.shareit.item.service.ItemServiceImpl;
 import ru.practicum.shareit.validation.ValidationGroups;
 
 import javax.validation.Valid;
@@ -22,7 +20,7 @@ import java.util.List;
 @RequestMapping("/items")
 @Validated
 public class ItemController {
-    private final ItemService itemService;
+    private final ItemServiceImpl itemService;
 
     @GetMapping
     public List<ItemDto> getOwnerAll(@RequestHeader("X-Sharer-User-Id") Integer ownerId) {
@@ -30,13 +28,9 @@ public class ItemController {
     }
 
     @GetMapping("/{id}")
-    public ItemDto getById(@PathVariable String id) {
-        try {
-            return itemService.findById(Integer.parseInt(id));
-        } catch (NumberFormatException e) {
-            log.error("The passed ID: {} is not Integer", id);
-            throw new MyValidationException(String.format("The passed ID: %s is not Integer", id));
-        }
+    public ItemDto getById(@PathVariable Integer id) {
+
+        return itemService.findById(id);
     }
 
     @GetMapping("/search")
@@ -46,21 +40,17 @@ public class ItemController {
     }
 
     @PostMapping
-    public Item create(@RequestBody @Validated(ValidationGroups.Create.class) ItemDto itemDto,
-                       @RequestHeader("X-Sharer-User-Id") Integer ownerId) {
+    public ItemDto create(@RequestBody @Validated(ValidationGroups.Create.class) ItemDto itemDto,
+                          @RequestHeader("X-Sharer-User-Id") Integer ownerId) {
 
         return itemService.create(itemDto, ownerId);
     }
 
     @PatchMapping("/{id}")
-    public Item update(@Valid @RequestBody ItemDto itemDto, @PathVariable String id,
-                       @RequestHeader("X-Sharer-User-Id") Integer ownerId) {
-        try {
-            return itemService.update(itemDto, ownerId, Integer.parseInt(id));
-        } catch (NumberFormatException e) {
-            log.error("The passed ID: {} is not Integer", id);
-            throw new MyValidationException(String.format("The passed ID: %s is not Integer", id));
-        }
+    public ItemDto update(@Valid @RequestBody ItemDto itemDto, @PathVariable Integer id,
+                          @RequestHeader("X-Sharer-User-Id") Integer ownerId) {
+
+        return itemService.update(itemDto, ownerId, id);
     }
