+++ src/main/java/ru/practicum/shareit/item/dto/ItemMapper.java
@@ -2,7 +2,7 @@ package ru.practicum.shareit.item.dto;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import ru.practicum.shareit.booking.service.BookingService;
+import ru.practicum.shareit.booking.service.BookingServiceImpl;
 import ru.practicum.shareit.item.model.Item;
 
 import java.util.ArrayList;
@@ -10,10 +10,10 @@ import java.util.List;
 
 @Component
 public class ItemMapper {
-    private final BookingService bookingService;
+    private final BookingServiceImpl bookingService;
 
     @Autowired
-    public ItemMapper(BookingService bookingService) {
+    public ItemMapper(BookingServiceImpl bookingService) {
         this.bookingService = bookingService;
     }
