+++ src/main/java/ru/practicum/shareit/user/controller/UserController.java
@@ -4,10 +4,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.exception.MyValidationException;
 import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.model.User;
-import ru.practicum.shareit.user.service.UserService;
+import ru.practicum.shareit.user.service.UserServiceImpl;
 import ru.practicum.shareit.validation.ValidationGroups;
 
 import javax.validation.Valid;
@@ -22,7 +21,7 @@ import java.util.List;
 @RequiredArgsConstructor
 @Validated
 public class UserController {
-    private final UserService userService;
+    private final UserServiceImpl userService;
 
     @GetMapping
     public List<UserDto> getAll() {
@@ -30,46 +29,32 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public UserDto getById(@PathVariable String id) {
-        try {
-            return userService.findById(Integer.parseInt(id));
-        } catch (NumberFormatException e) {
-            log.error("The passed ID: {} is not Integer", id);
-            throw new MyValidationException(String.format("The passed ID: %s is not Integer", id));
-        }
+    public UserDto getById(@PathVariable Integer id) {
+
+        return userService.findById(id);
     }
 
     @PostMapping
-    public User create(@RequestBody @Validated(ValidationGroups.Create.class) UserDto userDto) {
+    public UserDto create(@RequestBody @Validated(ValidationGroups.Create.class) UserDto userDto) {
 
         return userService.create(userDto);
     }
 
     @PutMapping
-    public User update(@Valid @RequestBody User user) {
+    public UserDto update(@Valid @RequestBody User user) {
 
         return userService.update(user);
     }
 
     @PatchMapping("/{id}")
-    public User updateByPatch(@Valid @RequestBody UserDto userDto, @PathVariable String id) {
+    public UserDto updateByPatch(@Valid @RequestBody UserDto userDto, @PathVariable Integer id) {
 
-        try {
-            return userService.updateByPatch(userDto, Integer.parseInt(id));
-        } catch (NumberFormatException e) {
-            log.error("The passed ID: {} is not Integer", id);
-            throw new MyValidationException(String.format("The passed ID: %s is not Integer", id));
-        }
-
+        return userService.updateByPatch(userDto, id);
     }
 
     @DeleteMapping("/{id}")
-    public User delete(@PathVariable String id) {
-        try {
-            return userService.delete(Integer.parseInt(id));
-        } catch (NumberFormatException e) {
-            log.error("The passed ID: {} is not Integer", id);
-            throw new MyValidationException(String.format("The passed ID: %s is not Integer", id));
-        }
+    public UserDto delete(@PathVariable Integer id) {
+
+        return userService.delete(id);
     }
 }
